async-coap-uri-macros: move error to separate file
